USB: Replace Train passthrough with compatibility modes
Testers reported that the passthrough mode doesn't always work for
various original Desha De Go! controllers. Rather than attempting to
cover all original controllers with a generic passthrough mode, this
patch replaces the setting with a set of compatibility modes that can be
extended to work for a range of different controller models.